+++ src/tracing/ipc/posix_shared_memory.cc
@@ -43,8 +43,8 @@ std::unique_ptr<PosixSharedMemory> PosixSharedMemory::Create(size_t size) {
   base::ScopedFile fd;
 #if PERFETTO_BUILDFLAG(PERFETTO_OS_ANDROID)
   bool is_memfd = false;
-  fd.reset(syscall(__NR_memfd_create, "perfetto_shmem",
-                   MFD_CLOEXEC | MFD_ALLOW_SEALING));
+  fd.reset(static_cast<int>(syscall(__NR_memfd_create, "perfetto_shmem",
+                                    MFD_CLOEXEC | MFD_ALLOW_SEALING)));
   is_memfd = !!fd;
 
   if (!fd) {
